Add ua node (#34)
Added UA node and miscellaneous fixes

+++ src/compact_genome.cpp
@@ -1,8 +1,5 @@
 #include "compact_genome.hpp"
 
-#include <range/v3/action/sort.hpp>
-#include <range/v3/action/unique.hpp>
-
 #include "dag.hpp"
 
 const CompactGenome* CompactGenome::Empty() {
@@ -160,7 +157,7 @@ size_t CompactGenome::ComputeHash(
   size_t result = 0;
   for (auto [pos, base] : mutations) {
     result = HashCombine(result, pos.value);
-    result = HashCombine(result, base);
+    result = HashCombine(result, static_cast<size_t>(base));
   }
   return result;
 }
